change `int` to `size_t`

+++ include/kernel.cuh
@@ -5,13 +5,13 @@
 #define WARP_SIZE 32;
 
 template<typename T>
-__global__ void printData(int n, T *a);
+__global__ void printData(size_t n, T *a);
 
 
-__global__ void convertFp32ToFp16(const int n, const float *in, half *out);
+__global__ void convertFp32ToFp16(const size_t n, const float *in, half *out);
 
 
-__global__ void comp_sddmm_gpu(const int M, const int N, const int K,
+__global__ void comp_sddmm_gpu(const size_t M, const size_t N, const size_t K,
                                const half *matrixA, const half *matrixB,
                                const float *matrixS,
                                float *matrixP);
